Health system and Damage
I've added a health system where there's a health bar and the player can be damaged. The projectiles hurt the player and the exploding enemies have a small delay before explosion. The stabbies still need work, but are getting done.  Still need to add the player doing damage. Sticking with just long range.