On Monday
I'll make a full level or two if I have the time and that will be turned in for the project! Two different guns are in. All enemies work, for the most part.  There's a jump power up, speed power up and two gun power ups. The bullets don't have assets and all are in 3d as spheres. The power ups are just white cubes and the zoomies are still some random sprite i found on google. I still need to make the main menu, but if I work all day monday, then that shouldn't be an issue. It's just adding a scene and some buttons then making sure that if a player gets past a certain boundary they win/move on to the next level. If we continue to work on this after the assignment is done, we can add more lol.